% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usa_dfp_map.R
\name{append_dfp_branding}
\alias{append_dfp_branding}
\title{Appends Data for Progress branding to a Data for Progress map.}
\usage{
append_dfp_branding(plot_path, annotation_text, font = "Montserrat")
}
\arguments{
\item{plot_path}{Path to the plot without Data for Progress branding.}

\item{annotation_text}{Text to include annotation.}

\item{font}{Font of choice with a limited selection to most base fonts. Defaults to Montserrat.}
}
\description{
Appends Data for Progress branding to a Data for Progress map.
}
