\name{DFP-internal}
\title{Internal DFP objects}
\alias{.skipOddValues}
\alias{.fuzzyDiscretization}
\alias{.fuzzyPatterns}
\alias{.calculateMembershipFunctions}
\alias{.discretizeExpressionValues}
\alias{.calculateFuzzyPatterns}
\alias{.calculateDiscriminantFuzzyPattern}
\alias{.plotGeneMF}
\alias{.computeMembershipLow}
\alias{.computeMembershipMedium}
\alias{.computeMembershipHigh}
\alias{.setValuesLow}
\alias{.setValuesMedium}
\alias{.setValuesHigh}
\description{Internal DFP objects.}
\details{These are not to be called by the user.}
\keyword{internal}