% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usa_dfp_map.R
\name{fixup}
\alias{fixup}
\title{Adjusts coordinates of Alaska and Hawaii.}
\usage{
fixup(usa, alaskaFix, hawaiiFix)
}
\arguments{
\item{usa}{}

\item{alaskaFix}{}
}
\description{
Adjusts coordinates of Alaska and Hawaii.
}
