% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_graph.R
\name{issue_analysis}
\alias{issue_analysis}
\title{Plot informative analytic data on the Twitter search.}
\usage{
issue_analysis(tweet_database, issue_terms)
}
\arguments{
\item{tweet_database}{Tweet database. Must have the columns "last_name", "party", "state" and "tweet".}

\item{issue_terms}{Search term or terms. Takes input as a vector.}
}
\description{
Plot informative analytic data on the Twitter search.
}
