% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_graph.R
\name{issue_bargraph}
\alias{issue_bargraph}
\title{Plot Data for Progress specified tweet graphs.}
\usage{
issue_bargraph(issue_terms, tweet_database, scale_by_total = FALSE,
  save_local = FALSE)
}
\arguments{
\item{issue_terms}{Search term or terms. Takes input as a vector.}

\item{tweet_database}{Tweet database. Must have the columns "last_name", "party", "state" and "tweet".}

\item{scale_by_total}{Scale your graph by all tweets by the Senator. Defaults to FALSE, which outputs tweet counts.}

\item{save_local}{Save the graph to your working directory. Defaults to FALSE.}
}
\description{
Plot Data for Progress specified tweet graphs.
}
