% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usa_dfp_map.R
\name{load_census_fips}
\alias{load_census_fips}
\title{Load U.S. Census FIPS Information from U.S Census website and returns a dataframe.}
\usage{
load_census_fips()
}
\description{
Load U.S. Census FIPS Information from U.S Census website and returns a dataframe.
}
