% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usa_dfp_map.R
\name{prelim_fix}
\alias{prelim_fix}
\title{Subfunction that adjusts specific coordinates.}
\usage{
prelim_fix(object, params)
}
\arguments{
\item{object}{}

\item{params}{}
}
\description{
Subfunction that adjusts specific coordinates.
}
