\name{show-methods}
\docType{methods}
\alias{show-methods}
\title{ Prints the slots (attributes) of an ExpressionLevel object }
\description{
 Prints the slots (center and width) of an \code{"\linkS4class{ExpressionLevel}"} object.
}
\section{Methods}{
\describe{

\item{object = "ExpressionLevel"}{ See \code{"\linkS4class{ExpressionLevel}"}. }
}}
\keyword{methods}