% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_graph.R
\name{twitter_analysis}
\alias{twitter_analysis}
\title{Plot informative analytic data on the Twitter search.}
\usage{
twitter_analysis(issue_terms, tweet_database)
}
\arguments{
\item{issue_terms}{Search term or terms. Takes input as a vector.}

\item{tweet_database}{Tweet database. Must have the columns "last_name", "party", "state" and "tweet".}
}
\description{
Plot informative analytic data on the Twitter search.
}
