% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usa_dfp_map.R
\name{usa_dfp_map}
\alias{usa_dfp_map}
\title{Plot Data for Progress specified USA map.}
\usage{
usa_dfp_map(states_data, color_value, alpha_value = NULL, title = NULL,
  subtitle = NULL, font = "Montserrat", color_scheme = "orange_to_green",
  limits = c(0.2, 0.8), save_plot = TRUE, output_folder = NULL)
}
\arguments{
\item{states_data}{State data frame that contains column entitle "abbr".}

\item{color_value}{Column name in the data frame for color fill value. Must be input as a character.}

\item{alpha_value}{Column name in the data frame for alpha transparency value. Must be input as a character. Defaults to NULL.}

\item{title}{Map title name. Defaults to NULL.}

\item{subtitle}{Map subtitle name. Defaults to NULL.}

\item{font}{Font of choice with a limited selection to most base fonts. Defaults to Montserrat.}

\item{color_scheme}{Map fill color scheme. Defaults to "orange_to_green". Also supports "brown_to_purple", "teal_to_pink", "plum_to_gold" and "red_to_blue".}

\item{limits}{Range of the color scale. Defaults to c(0.2, 0.8) for most polling data.}

\item{save_plot}{Saves plot if TRUE. Defaults to FALSE, which only displays the plot.}

\item{output_folder}{Output folder destination. Defaults to NULL, which outputs in the same folder.}
}
\description{
Plot Data for Progress specified USA map.
}
